+++ blurring_util.py
@@ -1,0 +1,68 @@
+#!/usr/bin/python
+# Helper functions for creating blur masks on image frames.
+
+# -*- coding: utf-8 -*-
+# pylint: disable=C0103
+# pylint: disable=E1101
+import numpy as np
+import cv2
+
+# Apply a Gaussian blur mask to a given image.
+def anonymize_face_gaussian(image, factor=3.0):
+	# Determine the size of the blurring kernel based on the dimensions of the image
+	(h, w) = image.shape[:2]
+	kW = int(w / factor)
+	kH = int(h / factor)
+
+	# Ensure the width of the kernel is odd
+	if kW % 2 == 0:
+		kW -= 1
+
+	# Ensure the height of the kernel is odd
+	if kH % 2 == 0:
+		kH -= 1
+
+	# Apply a Gaussian blur to the input image using our computed kernel size
+	return cv2.GaussianBlur(image, (kW, kH), 0)
+
+# Apply a pixelated blur mask to a given image.
+def anonymize_face_pixelate(image, blocks=3):
+	# Divide the input image into NxN blocks
+	(h, w) = image.shape[:2]
+	xSteps = np.linspace(0, w, blocks + 1, dtype="int")
+	ySteps = np.linspace(0, h, blocks + 1, dtype="int")
+
+	# Loop over the blocks in both the x and y direction
+	for i in range(1, len(ySteps)):
+		for j in range(1, len(xSteps)):
+			# Compute the starting and ending (x, y)-coordinates For the current block
+			startX = xSteps[j - 1]
+			startY = ySteps[i - 1]
+			endX = xSteps[j]
+			endY = ySteps[i]
+
+			# Extract the ROI using NumPy array slicing, compute the mean of the ROI, and 
+            # then draw a rectangle with the mean RGB values over the ROI in the original image
+			roi = image[startY:endY, startX:endX]
+			(B, G, R) = [int(x) for x in cv2.mean(roi)[:3]]
+			cv2.rectangle(image, (startX, startY), (endX, endY),
+				(B, G, R), -1)
+
+	# Return the pixelated blurred image
+	return image
+
+# Calculate the number of blocks to be used when creating the blurring mask.
+# This is based on the dimensions of the video frame. We take the maximum of the default
+# number of blocks and the calculated number of blocks to ensure that we use enough blocks
+# for an effective blur.
+def calculate_num_blocks_for_pixelated_blur(x1, y1, x2, y2, block_scaling_factor, default_num_blocks):
+    return max(int(max(abs(y1-y2), abs(x1-x2))*block_scaling_factor), default_num_blocks)
+
+# Create a blur mask for a given region of interest and merge it into a video frame.
+def create_and_merge_pixelated_blur_mask(frame, x1, y1, x2, y2, block_scaling_factor, default_num_blocks):
+    face = frame[y1:y2, x1:x2]
+    # Apply a pixelated blur on the object
+    face = anonymize_face_pixelate(face, calculate_num_blocks_for_pixelated_blur(x1, y1, x2, y2, block_scaling_factor, default_num_blocks))
+    # Merge the blurred object to our final image
+    frame[y1:y2, x1:x2] = face
+    return frame
