+++ file_util.py
@@ -1,0 +1,6 @@
+#!/usr/bin/python
+# Helper functions for validating and working with video files
+
+# -*- coding: utf-8 -*-
+# pylint: disable=C0103
+# pylint: disable=E1101
